# GameDebugClient/src/api/game.swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Rest Adventure - Game API",
    "version": "0.0.0-alpha.0.233+2a1139010fada7f6864f85e7bd96490563a587b1"
  },
  "servers": [
    {
      "url": "https://localhost:7056"
    }
  ],
  "paths": {
...
    "/game/content/locations/{locationId}": {
      "get": {
        "tags": [
          "Game"
        ],
        "summary": "Get location",
        "operationId": "GameContent_GetLocation",
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
...
    "/game/team/characters/{characterGuid}/jobs/harvestables": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get harvestables",
        "operationId": "JobsHarvest_GetHarvestables",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HarvestableEntity"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/team/characters/{characterGuid}/jobs/harvestables/{entityGuid}/{harvestName}": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Harvest",
        "operationId": "JobsHarvest_Harvest",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "entityGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          },
          {
            "name": "harvestName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/team/characters/{characterGuid}/locations": {
      "get": {
        "tags": [
          "Locations"
        ],
        "summary": "Get accessible locations",
        "operationId": "Locations_GetAccessibleLocations",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationWithAccess"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/team/characters/{characterGuid}/locations/{locationGuid}": {
      "post": {
        "tags": [
          "Locations"
        ],
        "summary": "Move to location",
        "operationId": "Locations_MoveToLocation",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "locationGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LocationMinimal": {
        "type": "object",
        "description": "Map location minimal information\n            ",
        "additionalProperties": false,
        "required": [
          "id",
          "area",
          "positionX",
          "positionY"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the location\n            ",
            "format": "guid",
            "minLength": 1
          },
          "area": {
            "description": "The area associated with the location\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/MapArea"
              }
            ]
          },
          "positionX": {
            "type": "integer",
            "description": "The X coordinate associated with the location\n            ",
            "format": "int32"
          },
          "positionY": {
            "type": "integer",
            "description": "The Y coordinate associated with the location\n            ",
            "format": "int32"
          }
        }
      },
      "MapArea": {
        "type": "object",
        "description": "Map area\n            ",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "level"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the area\n            ",
            "format": "guid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the area\n            ",
            "minLength": 1
          },
          "level": {
            "type": "integer",
            "description": "The level of the area\n            ",
            "format": "int32"
          }
        }
      },
...
      "ItemMinimal": {
        "type": "object",
        "description": "Item (minimal)\n            ",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "weight"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the item\n            ",
            "format": "guid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the item\n            ",
            "minLength": 1
          },
          "weight": {
            "type": "integer",
            "description": "The weight of the item\n            ",
            "format": "int32"
          }
        }
      },
...
      "JobMinimal": {
        "type": "object",
        "description": "Job (minimal)\n            ",
        "additionalProperties": false,
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the job\n            ",
            "format": "guid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the job\n            ",
            "minLength": 1
          }
        }
      },
...
      "HarvestableEntityHarvestMinimal": {
        "type": "object",
        "description": "Harvestable entity harvest (minimal\n            ",
        "additionalProperties": false,
        "required": [
          "job",
          "name",
          "level",
          "targets",
          "expectedHarvest",
          "expectedExperience"
        ],
        "properties": {
          "job": {
            "description": "The job providing the harvest\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/JobMinimal"
              }
            ]
          },
          "name": {
            "type": "string",
            "description": "The name of the harvest\n            ",
            "minLength": 1
          },
          "level": {
            "type": "integer",
            "description": "The level of the harvest\n            ",
            "format": "int32"
          },
          "targets": {
            "type": "array",
            "description": "The targets compatible with the harvest\n            ",
            "items": {
              "$ref": "#/components/schemas/StaticObject"
            }
          },
          "expectedHarvest": {
            "type": "array",
            "description": "The expected result of the harvest\n            ",
            "items": {
              "$ref": "#/components/schemas/ItemStack"
            }
          },
          "expectedExperience": {
            "type": "integer",
            "description": "The expected experience gain from the harvest\n            ",
            "format": "int32"
          }
        }
      },
      "StaticObject": {
        "type": "object",
        "description": "Static object\n            ",
        "additionalProperties": false,
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the static object\n            ",
            "format": "guid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the static object\n            ",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "description": "The description of the static object\n            ",
            "nullable": true
          }
        }
      },
      "ItemStack": {
        "type": "object",
        "description": "Item stack\n            ",
        "additionalProperties": false,
        "required": [
          "item",
          "count"
        ],
        "properties": {
          "item": {
            "description": "The item instance representing this stack\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/ItemMinimal"
              }
            ]
          },
          "count": {
            "type": "integer",
            "description": "The number of instances in this stack\n            ",
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "EntityMinimal": {
        "type": "object",
        "description": "Entity (minimal)\n            ",
        "additionalProperties": false,
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the entity\n            ",
            "format": "guid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the entity\n            ",
            "minLength": 1
          }
        }
      },
...
      "Location": {
        "allOf": [
          {
            "$ref": "#/components/schemas/LocationMinimal"
          },
          {
            "type": "object",
            "description": "Map location\n            ",
            "additionalProperties": false,
            "required": [
              "discovered",
              "connectedLocations"
            ],
            "properties": {
              "discovered": {
                "type": "boolean",
                "description": "Has this location been discovered by the player.\nIf false, the connected locations will be hidden.\n            "
              },
              "connectedLocations": {
                "type": "array",
                "description": "The locations connected to this one\n            ",
                "items": {
                  "$ref": "#/components/schemas/LocationMinimal"
                }
              }
            }
          }
        ]
      },
...
      "HarvestableEntity": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EntityMinimal"
          },
          {
            "type": "object",
            "description": "Harvestable entity\n            ",
            "additionalProperties": false,
            "required": [
              "harvests"
            ],
            "properties": {
              "harvests": {
                "type": "array",
                "description": "The harvests available on the entity\n            ",
                "items": {
                  "$ref": "#/components/schemas/HarvestableEntityHarvest"
                }
              }
            }
          }
        ]
      },
      "HarvestableEntityHarvest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/HarvestableEntityHarvestMinimal"
          },
          {
            "type": "object",
            "description": "Harvestable entity harvest\n            ",
            "additionalProperties": false,
            "required": [
              "canHarvest"
            ],
            "properties": {
              "canHarvest": {
                "type": "boolean",
                "description": "Can the harvest be performed\n            "
              },
              "whyCannotHarvest": {
                "type": "string",
                "description": "Why cannot the harvest be performed\n            ",
                "nullable": true
              }
            }
          }
        ]
      },
      "LocationWithAccess": {
        "type": "object",
        "description": "Location with access\n            ",
        "additionalProperties": false,
        "required": [
          "location",
          "isAccessible"
        ],
        "properties": {
          "location": {
            "description": "The location\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/LocationMinimal"
              }
            ]
          },
          "isAccessible": {
            "type": "boolean",
            "description": "Is the location accessible\n            "
          },
          "whyIsNotAccessible": {
            "type": "string",
            "description": "Why is the location not accessible\n            ",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
...
}